% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crt_buff.R
\name{crt_buff}
\alias{crt_buff}
\title{Reassign points based on distance to cluster boundaries}
\usage{
crt_buff(points, poly, distance = 50, trt, trt_value, id)
}
\arguments{
\item{points}{simple feature points locations of data}

\item{poly}{simple feature polygons of cluster boundaries}

\item{distance}{distance of buffer for cluster in metres}

\item{trt}{names of the intervention of treatment variable}

\item{trt_value}{value of the treatment to to buffer for}

\item{id}{name of id variable}
}
\value{
spatial points file with new ids and treatment variables
}
\description{
This function takes simple features points and polygon file.
It reassigns the points file to the intervention and id number of the cluster.
}
